+++ Project_Blind/Assets/Scripts/Core/StateMachineBehaviour/Player/MeleeAttackCombo4SMB.cs
@@ -34,7 +34,7 @@ namespace Blind
             {
                 if (_monoBehaviour.isJump)
                 {
-                    _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * _monoBehaviour.GetFacing());
+                    _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * (float)_monoBehaviour.GetFacing());
                 }
                 _monoBehaviour.enableAttack();
             }
@@ -59,7 +59,7 @@ namespace Blind
                 animator.speed = 1.0f;
                 _monoBehaviour._attack.DamageReset(_monoBehaviour.Data.powerAttackdamage);
                 _monoBehaviour.enableAttack();
-                _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * _monoBehaviour.GetFacing());
+                _monoBehaviour.AttackableMove(_monoBehaviour.Data.attackMove * (float)_monoBehaviour.GetFacing());
                 _monoBehaviour.CurrentWaveGauge -= 10;
                 _monoBehaviour.isPowerAttackEnd = false;
